# src/main/java/com/partnerhub/dto/ExternalProjectResponseDTO.java
package com.partnerhub.dto;

import java.time.Instant;

/**
 * Response payload representing an external project.
 */
public class ExternalProjectResponseDTO {

    private String id;
    private String name;
    private Instant createdAt;
    private Instant updatedAt;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Instant getCreatedAt() {
        return createdAt;
    }

    public void setCreatedAt(Instant createdAt) {
        this.createdAt = createdAt;
    }

    public Instant getUpdatedAt() {
        return updatedAt;
    }

    public void setUpdatedAt(Instant updatedAt) {
        this.updatedAt = updatedAt;
    }
}